+++ CHANGELOG.md
@@ -3,7 +3,10 @@ Not Released
 
 Added:
 
-* replace - it uses a gsub method and global replace
+* __ - first function call should have all arguments g(___2, __)(1, 3),
+  this means that it doesn't support g(__, 2)(1)(3), g(__, 2)(1, 3)
+  and g(__, 2)(__, 3)(1)
+* replace - it uses a gsub method and global replacement
 
 Release 0.5.0
 ---------------
@@ -30,7 +33,8 @@ Added:
 * ap
 * chain
 * complement
-* empty - uses x.class.new for defining new empty object if empty method does not exist
+* empty - uses x.class.new for defining new empty object if empty method
+  does not exist
 * from_pairs
 * is
 * of
@@ -60,7 +64,8 @@ Added:
 * binary
 * path
 * unary
-* uniq_with - first argument is a function with 1 arity which transforms each list element and applies uniq function to results
+* uniq_with - first argument is a function with 1 arity which transforms
+  each list element and applies uniq function to results
 * (pending) keys_in
 * (pending) values_in