+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Added:
 * __ - first function call should have all arguments g(___2, __)(1, 3), this means that it doesn't support g(__, 2)(1)(3), g(__, 2)(1, 3) and g(__, 2)(__, 3)(1)
 * bind
 * cond
+* prop_or
 * replace - it uses a gsub method and global replacement
 
 Release 0.5.0