+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Not Released
 ---------------
 
+Release 0.6.0
+
 Added:
 
 * F